fix(sync): fix ReconcileLocks race + move toFailedCycleOutcome to core:data
ConcurrentHashMap.getOrPut is non-atomic — two threads on Dispatchers.IO
calling withLock/withAnnotationLock for the same key simultaneously can each
create a distinct Mutex() and hold it independently, bypassing the
per-resource serialisation guarantee. Replace with computeIfAbsent which is
atomic at the ConcurrentHashMap level.

core:sync also had an inverted dependency on core:sources (for
AnnotationSyncException) because toFailedCycleOutcome lived in CycleOutcome.kt.
Move the classifier extension to CycleOutcomeMapper.kt in core:data, which
legitimately bridges both modules. Removes :core:sources from core:sync's
dependency graph.
