+++ core/data/src/main/kotlin/com/riffle/core/data/AnnotationLiveSync.kt
@@ -4,7 +4,7 @@ import com.riffle.core.domain.AnnotationSyncTarget
 import com.riffle.core.domain.DeviceIdStore
 import com.riffle.core.sync.AnnotationSyncStatusStore
 import com.riffle.core.sync.CycleOutcome
-import com.riffle.core.sync.toFailedCycleOutcome
+import com.riffle.core.data.toFailedCycleOutcome
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
