+++ core/data/src/main/kotlin/com/riffle/core/data/AnnotationSweep.kt
@@ -6,7 +6,7 @@ import com.riffle.core.sync.AnnotationSyncStatusStore
 import com.riffle.core.sync.CycleOutcome
 import com.riffle.core.sync.DirtyAnnotationLedger
 import com.riffle.core.sync.ReconcileLocks
-import com.riffle.core.sync.toFailedCycleOutcome
+import com.riffle.core.data.toFailedCycleOutcome
 import com.riffle.core.domain.DeviceIdStore
 import com.riffle.core.domain.DeviceLabelResolver
 import com.riffle.core.models.AnnotationFileHeader
