+++ core/sync/build.gradle.kts
@@ -6,7 +6,6 @@ dependencies {
     implementation(project(":core:domain"))
     implementation(project(":core:models"))
     implementation(project(":core:catalog"))
-    implementation(project(":core:sources"))
     implementation(libs.kotlinx.coroutines.core)
     implementation("javax.inject:javax.inject:1")
