+++ core/sync/src/main/kotlin/com/riffle/core/sync/CycleOutcome.kt
@@ -1,7 +1,5 @@
 package com.riffle.core.sync
 
-import com.riffle.core.sources.webdav.AnnotationSyncException
-
 /**
  * The outcome of one annotation-sync cycle (live push or sweep). The store keeps only the latest;
  * persisting across process death is rejected by ADR 0036.
@@ -19,13 +17,4 @@ sealed class CycleOutcome {
         data class Server(override val atMs: Long, val code: Int) : Failed(atMs)
         data class Unknown(override val atMs: Long, val message: String?) : Failed(atMs)
     }
-}
-
-/** Classify a thrown exception into a [CycleOutcome.Failed] subtype. */
-fun Throwable.toFailedCycleOutcome(at: Long): CycleOutcome.Failed = when (this) {
-    is AnnotationSyncException.AuthFailed -> CycleOutcome.Failed.Auth(at, code)
-    is AnnotationSyncException.HttpFailure -> CycleOutcome.Failed.Server(at, code)
-    is AnnotationSyncException.NetworkError -> CycleOutcome.Failed.Network(at, message)
-    is AnnotationSyncException.TlsError -> CycleOutcome.Failed.Tls(at, message)
-    else -> CycleOutcome.Failed.Unknown(at, message)
 }
